git-svn-id: https://svn.code.sf.net/p/rathena/svn/branches/stable@595 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/db.c
@@ -442,7 +442,7 @@ void db_foreach(struct dbt *table,int (*func)(void*,void*,va_list),...)
 		sp=0;
 		while(1){
 			if (!p->data) {
-				printf("Warning: no data for key %d in db_foreach (db.c) !\n",p->key);
+				printf("Warning: no data for key %d in db_foreach (db.c) !\n",(int)p->key);
 			} else {
 				func(p->key, p->data, ap);
 			}
